+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
 
-## 0.1
+## 0.2 - 21/09/2021
+- Add landmark detection.
+- Modify `README` file to have a comprehensive documentation.
+- Rename `analyze` method to `detect`.
+- Rename `detectFaces` to `faceDetection`.
+- Rename `detectImageText` to `imageTextDetection`.
+- Rename `detectImageProperties` to `imagePropertiesDetection`.
+- Rename `detectSafeSearch` to `safeSearchDetection`.
+
+## 0.1 - - 20/09/2021
 - Initial release.